Global Changing! new class Added: CompensationOptionNamesEnum DialogSettings - Port number removed, accuracy settings changed to coordinates accuracy  and corners accuracy CmmApprox - PortNumber removed;TriangulationINAccuracy, CoordinatesAccuracy and CornersAccuracy were Added ObjectApprox - in constructor objMath = nullptr; CompensationOption Added GeomObjectApprox - virtual function Triangulation() Added; objectApproxName removed in destructor; unnecessary pointer removed from method FindHeight() CircleGeometric - new constructor(Line, Radius) Added DialogObjectProperties - saving coordinates in file by object Name; Compensation option Added ...e.t.c.

+++ include/ApproximationCore/GlobalFunctions.h
@@ -13,10 +13,10 @@ void			GaussMethod(double	**coefficients,
 							int		dimension, 
 							double	*result);
 
-void			KramerMethod(	double** gaussCoefficients,
-								double* gaussFreeCoefficients,
-								double* gaussResult,
-								int D);
+void			KramerMethod(	double** gaussCoefficients, 
+								double* gaussFreeCoefficients, 
+								double* gaussResult, 
+								double D);
 
 PointGeometric	TransferPointToNewCoordinateSystem(	PointGeometric	point, 
 													PointGeometric	pointCenterNewCoordinateSystem,
